move binary reading to readers module
