+++ simulariumio/readers/__init__.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from .simularium_binary_reader import SimulariumBinaryReader  # noqa: F401
+from .binary_info import BinaryFileData, BinaryBlockInfo  # noqa: F401
