+++ simulariumio/readers/binary_info.py
