+++ simulariumio/tests/plot_readers/test_scatter_plot_reader.py
