+++ simulariumio/tests/readers/test_binary_reader.py
